Re-added Parcour edit menu
